requirements.txt
flask
flask-cors
yfinance
finnhub-python
requests
transformers
torch
pandas
scipy
numpy
